// src/types/video.ts
export interface VideoItem {
  id: string;
  title: string;
  uploaderName: string;
  uploaderId?: string;
  uploaderUrl?: string;
  uploaderAvatarUrl?: string;
  uploadDate?: string;
  viewCount: number;
  duration: number; // in seconds
  thumbnailUrl: string;
  isLive?: boolean;
  isUpcoming?: boolean;
  streamType?: 'VIDEO_STREAM' | 'LIVE_STREAM' | 'SHORTS' | 'AUDIO_STREAM';
  progress?: number; // watch progress in seconds
  localUri?: string; // For downloaded offline videos
}

export interface StreamItem {
  url: string;
  quality: string; // e.g. "1080p", "720p", "Auto"
  format: string; // e.g. "mp4", "m3u8", "webm"
  isAdaptive?: boolean;
  bitrate?: number;
  height?: number;
  mimeType?: string;
  codec?: string;
  audioOnly?: boolean;
  headers?: Record<string, string>;
}

export interface SubtitleItem {
  languageName: string;
  languageCode: string;
  url: string;
  isAutoGenerated?: boolean;
}

export interface StreamBundle {
  expiresAt?: number;
  storyboards?: Storyboard[];
  chapters?: VideoChapter[];
  videoId: string;
  title: string;
  uploaderName: string;
  uploaderId?: string;
  uploaderUrl?: string;
  uploaderAvatarUrl?: string;
  uploaderSubscriberCount?: string | number;
  description?: string;
  viewCount: number;
  uploadDate?: string;
  thumbnailUrl: string;
  isLive: boolean;
  isUpcoming: boolean;
  scheduledStartTime?: string | null;
  videoStreams: StreamItem[];
  audioStreams: StreamItem[];
  qualityStreams?: StreamItem[];
  bestAudioStreamUrl?: string;
  subtitles: SubtitleItem[];
  relatedVideos: VideoItem[];
  dashManifestUrl?: string;
  hlsManifestUrl?: string;
  hlsManifestHeaders?: Record<string, string>;
  dashManifestHeaders?: Record<string, string>;
}

export interface Storyboard { urls?: string[]; templateUrl: string; width: number; height: number; count: number; columns: number; rows: number; intervalMs: number; }
export interface VideoChapter { title: string; startTime: number; thumbnailUrl?: string; }